pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-analysis-sdk"
version = "0.1.0"
description = "Analyze, score, and optimize LLM prompts for clarity, efficiency, and token cost"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"

authors = [
  { name = "Your Name", email = "your@email.com" }
]

dependencies = [
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "typer>=0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.7",
]

[project.scripts]
promptlint = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["prompt_analysis", "cli"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]